requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.8.4
    # via fastapi-cache2
aiosignal==1.3.1
    # via aiohttp
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
async-timeout==4.0.2
    # via
    #   aiohttp
    #   redis
attrs==22.2.0
    # via aiohttp
azure-core==1.26.4
    # via azure-storage-blob
azure-storage-blob==12.16.0
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
click==8.1.3
    # via uvicorn
cryptography==40.0.2
    # via azure-storage-blob
duckdb==0.7.1
    # via -r requirements.in
fastapi==0.95.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
    #   fastapi-versionizer
    #   prometheus-fastapi-instrumentator
fastapi-cache2[redis]==0.2.1
    # via -r requirements.in
fastapi-versionizer==0.1.2
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements.in
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
    #   yarl
isodate==0.6.1
    # via azure-storage-blob
loguru==0.6.0
    # via -r requirements.in
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multimethod==1.9.1
    # via pandera
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.24.2
    # via
    #   pandas
    #   pandera
    #   pyarrow
packaging==23.0
    # via pandera
pandas==1.4.1
    # via
    #   -r requirements.in
    #   pandera
pandera==0.14.5
    # via -r requirements.in
pendulum==2.1.2
    # via fastapi-cache2
prometheus-client==0.16.0
    # via prometheus-fastapi-instrumentator
prometheus-fastapi-instrumentator==6.0.0
    # via -r requirements.in
pyarrow==11.0.0
    # via -r requirements.in
pycparser==2.21
    # via cffi
pydantic[dotenv]==1.10.7
    # via
    #   -r requirements.in
    #   fastapi
    #   pandera
python-dateutil==2.8.2
    # via
    #   pandas
    #   pendulum
python-dotenv==1.0.0
    # via pydantic
pytz==2022.7.1
    # via pandas
pytzdata==2020.1
    # via pendulum
redis==4.5.4
    # via
    #   -r requirements.in
    #   fastapi-cache2
requests[security]==2.28.2
    # via
    #   -r requirements.in
    #   azure-core
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via
    #   azure-core
    #   isodate
    #   python-dateutil
    #   tenacity
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.26.1
    # via
    #   fastapi
    #   fastapi-versionizer
tenacity==6.2.0
    # via -r requirements.in
typing-extensions==4.5.0
    # via
    #   azure-core
    #   azure-storage-blob
    #   pydantic
    #   typing-inspect
typing-inspect==0.8.0
    # via pandera
urllib3==1.26.15
    # via requests
uvicorn==0.21.0
    # via
    #   -r requirements.in
    #   fastapi-cache2
uvloop==0.17.0
    # via -r requirements.in
wrapt==1.15.0
    # via pandera
yarl==1.8.2
    # via aiohttp
